+++ demo/androidsdk/src/main/java/com/loginradius/androidsdk/activity/FacebookNativeActivity.java
@@ -104,11 +104,11 @@ public class FacebookNativeActivity extends AppCompatActivity {
             public void onSuccess(AccessTokenResponse data) {
                 accessToken = data;
                 accessToken.provider = "facebook";
-                accessToken.apikey = LoginRadiusSDK.getApiKey();
+
                 if(isRequired){
                     getRaasSchema();
                 }else{
-                    sendAccessToken(data.access_token);
+                    sendAccessToken(data.access_token,data.refresh_token);
                 }
             }
 
@@ -137,7 +137,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
             @Override
             public void onFailure(Throwable error, String errorcode) {
                 Log.i("lr_api_error",error.getMessage());
-                sendAccessToken(null);
+                sendAccessToken(null,null);
             }
         });
     }
@@ -152,7 +152,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
         if(containsRequired){
             getUserProfile();
         }else{
-            sendAccessToken(accessToken.access_token);
+            sendAccessToken(accessToken.access_token,accessToken.refresh_token);
         }
     }
 
@@ -171,7 +171,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
             @Override
             public void onFailure(Throwable error, String errorcode) {
                 Log.i("lr_api_error",error.getMessage());
-                sendAccessToken(null);
+                sendAccessToken(null,null);
             }
         });
     }
@@ -196,7 +196,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
                 if(value.getChildCount()>0){
                     setContentView(value);
                 }else{
-                    sendAccessToken(accessToken.access_token);
+                    sendAccessToken(accessToken.access_token,accessToken.refresh_token);
                 }
             }
 
@@ -277,7 +277,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
                         showEmailInfoDialog();
                         return;
                     }
-                    sendAccessToken(accessToken.access_token);
+                    sendAccessToken(accessToken.access_token,accessToken.refresh_token);
                 }
 
                 @Override
@@ -317,7 +317,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
         alert.setPositiveButton("OK", new OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
-                sendAccessToken(null);
+                sendAccessToken(null,null);
             }
         });
         AlertDialog dialog = alert.create();
@@ -399,7 +399,7 @@ public class FacebookNativeActivity extends AppCompatActivity {
         api.verifyOtp(queryParams, jsonObject, new AsyncHandler<LoginData>() {
             @Override
             public void onSuccess(LoginData data) {
-                sendAccessToken(data.getAccessToken());
+                sendAccessToken(data.getAccessToken(),data.getRefreshToken());
             }
 
             @Override
@@ -429,12 +429,14 @@ public class FacebookNativeActivity extends AppCompatActivity {
     }
 
 
-    public void sendAccessToken(String accessToken) {
+    public void sendAccessToken(String accessToken,String refreshToken) {
         AccessTokenResponse accesstoken = new AccessTokenResponse();
         accesstoken.access_token = accessToken;
+        accesstoken.refresh_token=refreshToken;
         accesstoken.provider = "facebook";
         Intent intent = new Intent();
         intent.putExtra("accesstoken", accessToken);
+        intent.putExtra("refreshtoken",refreshToken);
         intent.putExtra("provider", "facebook");
         setResult(2, intent);
         finish();//finishing activity
@@ -447,9 +449,11 @@ public class FacebookNativeActivity extends AppCompatActivity {
         if (requestCode == 2) {
             if (data != null) {
                 String token = data.getStringExtra("accesstoken");
+                String refreshtoken=data.getStringExtra("refreshtoken");
                 String provider = data.getStringExtra("provider");
                 Intent intent = new Intent();
                 intent.putExtra("accesstoken", token);
+                intent.putExtra("refreshtoken",refreshtoken);
                 intent.putExtra("provider", provider);
                 setResult(2, intent);
                 finish();//finishing activity