+++ demo/app/src/main/java/com/loginradius/demo/MainActivity.java
@@ -4,9 +4,10 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
+
 import android.view.View;
 import android.widget.Button;
-
+import com.loginradius.androidsdk.activity.VkontakteNativeActivity;
 import com.loginradius.androidsdk.helper.LoginRadiusSDK;
 import com.loginradius.androidsdk.helper.LoginRadiusSDK.Initialize;
 
@@ -59,10 +60,13 @@ public class MainActivity extends AppCompatActivity {
         btnlogin = (Button) findViewById(R.id.login_bt);
         btnregister = (Button) findViewById(R.id.signup_bt);
         btnforgotpassword = (Button) findViewById(R.id.forgot_bt);
+
         //set on Click listener
         btnlogin.setOnClickListener(listener);
         btnregister.setOnClickListener(listener);
         btnforgotpassword.setOnClickListener(listener);
 
 }
+
+
 }